+++ core/backgammongame.py
@@ -12,8 +12,7 @@ class backgammongame:
         self.__turno__ = self.__jugador1__
         self.__turno_finalizado__ = False
 
-# Funciones básicas:
-
+    # Funciones básicas:
     def mostrar_jugador1 (self):
         return self.__jugador1__
     
@@ -26,24 +25,86 @@ class backgammongame:
     def mostrar_tablero(self):
         return self.__board__
     
-#Funciones de dados
-
+    #Funciones de dados
     def tirar_dados(self):
-        self.__dados__.tirar_dados()
-        return self.__dados__.get_dados()
+        self.__dados__.tirar_dados() #Tira los dados
+        return self.__dados__.get_dados() #Me devuelve los números que salieron
     
+    def mover(self, origen: int, destino: int):
+        movimiento = abs(destino - origen)
+
+        if movimiento not in self.__dados__.get_dados(): #Verificamos si está disponible el movimiento
+            raise ValueError(f"Movimiento {movimiento} no está disponible en los dados {self.__dados__.get_dados()}")
+
+        self.__board__.mover_checker(origen, destino) #Movemos la ficha
+        self.__dados__.usar_dado(movimiento) #Sacamos el movimiento de los dados
+    
+    def reingresar_ficha(self, destino: int):
+        color = self.__turno__.obtener_color()
+        movimiento = destino + 1
+
+        if movimiento not in self.__dados__.get_dados(): #Verificamos si está disponible el movimiento
+            raise ValueError(f"Movimiento {movimiento} no está disponible en los dados {self.__dados__.get_dados()}")
+        
+        self.__board__.move_checker_banco(color, destino)
+        self.__dados__.usar_dado(movimiento)
+    
+    def sacar(self, origen: int): #CONSULTAR EN CLASE!!!
+        color = self.__turno__.obtener_color()
+
+        if color == "white":
+            fichas_fuera_cuadrante = any(
+                color in punto for i, punto in enumerate(self.__board__.mostrar_tablero()) if i < 18
+            )
+        else:  # black
+            fichas_fuera_cuadrante = any(
+                color in punto for i, punto in enumerate(self.__board__.mostrar_tablero()) if i > 5
+            )
+
+        if fichas_fuera_cuadrante:
+            raise ValueError(f"El jugador {color} no puede sacar fichas: todavía tiene piezas fuera de su cuadrante final")
+
+        # Calcular el valor necesario para sacar
+        movimiento = 24 - origen if color == "white" else origen + 1
+
+        # Validar dado
+        if movimiento not in self.__dados__.get_dados():
+            raise ValueError(f"No puedes sacar desde {origen}, el valor {movimiento} no está en los dados {self.__dados__.get_dados()}")
+
+        # Intentar sacar ficha
+        if self.__board__.sacar_ficha(color, origen):
+            self.__turno__.restar_ficha()
+            self.__dados__.usar_dado(movimiento)
+            return True
+        return False
+
+    #Funciones del turno
     def cambiar_turno(self):
         if self.__turno__ == self.__jugador1__:
-            self.__turno__ = self.__jugador2__
+            self.__dados__.limpiar_dados() #Elimina los dados del turno anterior
+            self.__turno__ = self.__jugador2__ #Cambia de jugador
         else:
-            self.__turno__ = self.__jugador1__
+            self.__dados__.limpiar_dados() #Elimina los dados del turno anterior
+            self.__turno__ = self.__jugador1__ #Cambia de jugador
+    
+    def finalizar_turno(self):
+        if not self.__dados__.hay_movimientos(): #Corrobora que no quedan movimientos posibles
+            self.cambiar_turno()
+            self.__turno_finalizado__ = True
+        else:
+            self.__turno_finalizado__ = False
+    
+    def estado_turno(self):
+        color = self.__turno__.obtener_color()
+        return self.__board__.estado_jugador(color)
 
+#Funciones para ganar
     def ganador(self):
-        for jugador in self.__jugadores__: 
-            if jugador.ganar():
-                self.__juego_terminado__ = True
-                return jugador.obtener_nombre()
+        for jugador in self.__jugadores__: #Revisa la lista de jugadores
+            if jugador.ganar(): #Corrobora condición de ganar
+                self.__juego_terminado__ = True 
+                return jugador.obtener_nombre() #Me devuelve el nombre del ganador
         return None
     
     def juego_terminado(self):
-        return self.__jugador1__.ganar() or self.__jugador2__.ganar()
+        return self.__jugador1__.ganar() or self.__jugador2__.ganar() #Finaliza el juego si cualquiera de los jugadores gana